IKC-22 new frontend - consumer groups deletion

+++ kouncil-frontend/src/app/consumers/consumer-groups/consumer-groups.component.ts
@@ -7,6 +7,7 @@ import {ArraySortPipe} from '../../util/array-sort.pipe';
 import {ConsumerGroupsService} from './consumer-groups.service';
 import {first} from 'rxjs/operators';
 import {Router} from '@angular/router';
+import {ConfirmService} from '../../confirm/confirm.service';
 
 @Component({
   selector: 'kafka-consumer-groups',
@@ -18,6 +19,7 @@ export class ConsumerGroupsComponent implements OnInit, OnDestroy {
               private progressBarService: ProgressBarService,
               private arraySortPipe: ArraySortPipe,
               private consumerGroupsService: ConsumerGroupsService,
+              private confirmService: ConfirmService,
               private router: Router) {
   }
 
@@ -93,20 +95,26 @@ export class ConsumerGroupsComponent implements OnInit, OnDestroy {
   }
 
   deleteConsumerGroup(value) {
-    this.progressBarService.setProgress(true);
-    this.consumerGroupsService.deleteConsumerGroup(value)
+    this.confirmService.openConfirmDialog('consumer group', value)
       .pipe(first())
-      .subscribe(data => {
-        this.loadConsumerGroups();
-      }, error => {
-        console.warn(error);
-        this.progressBarService.setProgress(false);
+      .subscribe((confirmed) => {
+        if (confirmed) {
+          this.progressBarService.setProgress(true);
+          this.consumerGroupsService.deleteConsumerGroup(value)
+            .pipe(first())
+            .subscribe(data => {
+              this.loadConsumerGroups();
+            }, error => {
+              console.warn(error);
+              this.progressBarService.setProgress(false);
+            });
+        }
       });
   }
 
   navigateToConsumerGroup(event): void {
     const element = event.event.target as HTMLElement;
-    if (event.type === 'click' && element.nodeName !== 'MAT-ICON') {
+    if (event.type === 'click' && element.nodeName !== 'MAT-ICON' && element.nodeName !== 'BUTTON') {
       this.router.navigate(['/consumer-groups/', event.row.groupId]);
     }
   }
